Move SWF type checks to response object
(suggested in @spulec review)
